+++ internal/config/inbox.go
@@ -2,4 +2,5 @@ package config
 
 type Inbox struct {
 	StorageAddress string `env:"INBOX_API_STORAGE_ADDRESS" envDefault:"localhost:11055"`
+	FeedAddress    string `env:"INBOX_API_FEED_ADDRESS" envDefault:"localhost:11066"`
 }