+++ internal/rest/server.go
@@ -27,14 +27,16 @@ type Server struct {
 	coreclient  *coresdk.Client
 	subclient   inboxapi.SubscriptionClient
 	settings    inboxapi.SettingsClient
+	feedClient  inboxapi.FeedClient
 }
 
-func NewServer(cfg config.REST, authStorage AuthStorage, cl *coresdk.Client, sc inboxapi.SubscriptionClient, settings inboxapi.SettingsClient) *Server {
+func NewServer(cfg config.REST, authStorage AuthStorage, cl *coresdk.Client, sc inboxapi.SubscriptionClient, settings inboxapi.SettingsClient, feedClient inboxapi.FeedClient) *Server {
 	srv := &Server{
 		authStorage: authStorage,
 		coreclient:  cl,
 		subclient:   sc,
 		settings:    settings,
+		feedClient:  feedClient,
 	}
 
 	handler := mux.NewRouter()
@@ -76,6 +78,7 @@ func NewServer(cfg config.REST, authStorage AuthStorage, cl *coresdk.Client, sc 
 	handler.HandleFunc("/feed", srv.getFeed).Methods(http.MethodGet).Name("get_feed")
 	handler.HandleFunc("/feed/mark-as-read", srv.markAsReadBatch).Methods(http.MethodPost).Name("mark_as_read_batch")
 	handler.HandleFunc("/feed/{id}/mark-as-read", srv.markFeedItemAsRead).Methods(http.MethodPost).Name("mark_feed_item_as_read")
+	handler.HandleFunc("/feed/{id}/archive", srv.markFeedItemAsAcrhived).Methods(http.MethodPost).Name("mark_feed_item_as_archived")
 
 	handler.HandleFunc("/notifications/settings", srv.storePushToken).Methods(http.MethodPost).Name("store_push_token")
 	handler.HandleFunc("/notifications/settings", srv.tokenExists).Methods(http.MethodGet).Name("push_token_exists")