+++ internal/rest/server_feed.go
@@ -1,32 +1,28 @@
 package rest
 
 import (
+	"context"
+	"encoding/json"
 	"net/http"
-	"time"
 
 	"github.com/google/uuid"
 	"github.com/gorilla/mux"
+	coredao "github.com/goverland-labs/core-web-sdk/dao"
+	coreproposal "github.com/goverland-labs/core-web-sdk/proposal"
+	"github.com/goverland-labs/inbox-api/protobuf/inboxapi"
 	"github.com/rs/zerolog/log"
-	"github.com/samber/lo"
+	"google.golang.org/protobuf/types/known/timestamppb"
 
 	"github.com/goverland-labs/inbox-web-api/internal/appctx"
-	"github.com/goverland-labs/inbox-web-api/internal/auth"
 	"github.com/goverland-labs/inbox-web-api/internal/entities/common"
+	"github.com/goverland-labs/inbox-web-api/internal/entities/dao"
 	"github.com/goverland-labs/inbox-web-api/internal/entities/feed"
+	"github.com/goverland-labs/inbox-web-api/internal/entities/proposal"
 	"github.com/goverland-labs/inbox-web-api/internal/helpers"
 	feedform "github.com/goverland-labs/inbox-web-api/internal/rest/forms/feed"
-	"github.com/goverland-labs/inbox-web-api/internal/rest/mock"
 	"github.com/goverland-labs/inbox-web-api/internal/rest/request"
 	"github.com/goverland-labs/inbox-web-api/internal/rest/response"
 )
-
-type readMark struct {
-	SessionID  uuid.UUID
-	FeedItemID uuid.UUID
-	ReadAt     time.Time
-}
-
-var readMarks = make([]readMark, 0)
 
 func (s *Server) getFeed(w http.ResponseWriter, r *http.Request) {
 	session, ok := appctx.ExtractUserSession(r.Context())
@@ -35,28 +31,29 @@ func (s *Server) getFeed(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	list := getSessionFeed(session)
-
-	f, _ := feedform.NewGetFeedForm().ParseAndValidate(r)
-	list = lo.Filter(list, func(item feed.Item, index int) bool {
-		if !f.Unread {
-			return true
-		}
-
-		return item.ReadAt == nil
-	})
-
-	list = helpers.WrapDAFeedItemsIpfsLinks(list)
-
 	offset, limit, err := request.ExtractPagination(r)
 	if err != nil {
 		response.SendError(w, http.StatusBadRequest, err.Error())
 		return
 	}
 
-	totalCount := len(list)
+	f, _ := feedform.NewGetFeedForm().ParseAndValidate(r)
 
-	list = lo.Slice(list, offset, offset+limit)
+	resp, err := s.feedClient.GetUserFeed(context.TODO(), &inboxapi.GetUserFeedRequest{
+		SubscriberId:    session.ID.String(),
+		IncludeRead:     f.Unread,
+		IncludeArchived: f.Arhived,
+		Limit:           uint32(limit),
+		Offset:          uint32(offset),
+	})
+	if err != nil {
+		response.SendError(w, http.StatusInternalServerError, err.Error())
+		return
+	}
+
+	list := helpers.WrapFeedItemsIpfsLinks(convertInboxFeedListToInternal(resp.GetList()))
+	totalCount := int(resp.GetTotalCount())
+
 	log.Info().
 		Str("route", mux.CurrentRoute(r).GetName()).
 		Int("count", len(list)).
@@ -64,6 +61,7 @@ func (s *Server) getFeed(w http.ResponseWriter, r *http.Request) {
 		Msg("route execution")
 
 	response.AddPaginationHeaders(w, r, offset, limit, totalCount)
+	response.AddUnreadHeader(w, r, int(resp.UnreadCount))
 	response.SendJSON(w, http.StatusOK, &list)
 }
 
@@ -80,18 +78,40 @@ func (s *Server) markFeedItemAsRead(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	item, exist := mock.GetFeedItem(f.ID)
-	if !exist {
-		response.HandleError(response.NewNotFoundError(), w)
+	_, err := s.feedClient.MarkAsRead(context.TODO(), &inboxapi.MarkAsReadRequest{
+		SubscriberId: session.ID.String(),
+		Ids:          []string{f.ID.String()},
+	})
+
+	if err != nil {
+		response.HandleError(response.ResolveError(err), w)
 		return
 	}
 
-	if readAt(session, item) == nil {
-		readMarks = append(readMarks, readMark{
-			SessionID:  session.ID,
-			FeedItemID: item.ID,
-			ReadAt:     time.Now(),
-		})
+	response.SendEmpty(w, http.StatusOK)
+}
+
+func (s *Server) markFeedItemAsAcrhived(w http.ResponseWriter, r *http.Request) {
+	session, ok := appctx.ExtractUserSession(r.Context())
+	if !ok {
+		w.WriteHeader(http.StatusUnauthorized)
+		return
+	}
+
+	f, verr := feedform.NewMarkItemAsReadForm().ParseAndValidate(r)
+	if verr != nil {
+		response.HandleError(verr, w)
+		return
+	}
+
+	_, err := s.feedClient.MarkAsArchived(context.TODO(), &inboxapi.MarkAsArchivedRequest{
+		SubscriberId: session.ID.String(),
+		Ids:          []string{f.ID.String()},
+	})
+
+	if err != nil {
+		response.HandleError(response.ResolveError(err), w)
+		return
 	}
 
 	response.SendEmpty(w, http.StatusOK)
@@ -110,97 +130,95 @@ func (s *Server) markAsReadBatch(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	var list []feed.Item
-	var before = time.Now()
-
-	if len(f.IDs) > 0 {
-		list = getSessionFeedByID(session, f.IDs...)
-	} else {
-		list = getSessionFeed(session)
-		if f.Before != nil {
-			before = *f.Before
-		}
+	ids := make([]string, 0, len(f.IDs))
+	for _, id := range f.IDs {
+		ids = append(ids, id.String())
 	}
 
-	lo.ForEach(list, func(item feed.Item, index int) {
-		if item.ReadAt != nil {
-			return
-		}
-
-		if item.CreatedAt.After(before) {
-			return
-		}
-
-		readMarks = append(readMarks, readMark{
-			SessionID:  session.ID,
-			FeedItemID: item.ID,
-			ReadAt:     time.Now(),
-		})
-	})
-}
-
-func getSessionFeed(session auth.Session) []feed.Item {
-	subs := subscriptionsStorage[session.ID]
-	list := lo.Filter(mock.Feed, func(item feed.Item, _ int) bool {
-		var daoID uuid.UUID
-		if item.DAO != nil {
-			daoID = item.DAO.ID
-		} else if item.Proposal != nil {
-			daoID = item.Proposal.DAO.ID
-		}
-
-		for _, sub := range subs {
-			if sub.DAO == nil {
-				continue
-			}
-
-			if sub.DAO.ID == daoID {
-				return true
-			}
-		}
-
-		return false
-	})
-
-	return enrichReadMarks(session, list)
-}
-
-func getSessionFeedByID(session auth.Session, id ...uuid.UUID) []feed.Item {
-	list := getSessionFeed(session)
-	list = lo.Filter(list, func(item feed.Item, index int) bool {
-		for i := range id {
-			if item.ID == id[i] {
-				return true
-			}
-		}
-
-		return false
-	})
-
-	return list
-}
-
-func enrichReadMarks(session auth.Session, list []feed.Item) []feed.Item {
-	lo.ForEach(list, func(item feed.Item, index int) {
-		mark := readAt(session, item)
-		if mark != nil {
-			list[index].ReadAt = common.NewTime(*mark)
-		}
-	})
-
-	return list
-}
-
-func readAt(session auth.Session, item feed.Item) *time.Time {
-	for _, m := range readMarks {
-		if m.SessionID != session.ID {
-			continue
-		}
-
-		if item.ID == m.FeedItemID {
-			return helpers.Ptr(m.ReadAt)
-		}
+	var before *timestamppb.Timestamp
+	if f.Before != nil {
+		before = timestamppb.New(*f.Before)
 	}
 
-	return nil
+	_, err := s.feedClient.MarkAsRead(context.TODO(), &inboxapi.MarkAsReadRequest{
+		SubscriberId: session.ID.String(),
+		Ids:          ids,
+		Before:       before,
+	})
+
+	if err != nil {
+		response.HandleError(response.ResolveError(err), w)
+		return
+	}
+
+	response.SendEmpty(w, http.StatusOK)
+}
+
+func convertInboxFeedListToInternal(list []*inboxapi.FeedItem) []feed.Item {
+	converted := make([]feed.Item, 0, len(list))
+
+	for _, item := range list {
+		converted = append(converted, convertInboxFeedItemToInternal(item))
+	}
+
+	return converted
+}
+
+func convertInboxFeedItemToInternal(item *inboxapi.FeedItem) feed.Item {
+	var daoItem *dao.DAO
+	var proposalItem *proposal.Proposal
+
+	switch item.Type {
+	case "dao":
+		var daoSnapshot *coredao.Dao
+		if err := json.Unmarshal(item.GetSnapshot(), &daoSnapshot); err != nil {
+			log.Error().Err(err).Str("feed_id", item.GetId()).Msg("unable to unmarshal dao snapshot")
+		}
+
+		daoItem = helpers.Ptr(helpers.WrapDAOIpfsLinks(convertCoreDaoToInternal(daoSnapshot)))
+	case "proposal":
+		var proposalSnapshot *coreproposal.Proposal
+		if err := json.Unmarshal(item.GetSnapshot(), &proposalSnapshot); err != nil {
+			log.Error().Err(err).Str("feed_id", item.GetId()).Msg("unable to unmarshal proposal snapshot")
+		}
+
+		d := coredao.Dao{} // FIXME
+		proposalItem = helpers.Ptr(helpers.WrapProposalIpfsLinks(convertProposalToInternal(proposalSnapshot, &d)))
+	}
+
+	feedID, err := uuid.Parse(item.GetId())
+	if err != nil {
+		log.Error().Err(err).Str("id", item.GetId()).Msg("unable to parse feed id")
+	}
+
+	daoID, err := uuid.Parse(item.GetDaoId())
+	if err != nil {
+		log.Error().Err(err).Str("id", item.GetDaoId()).Msg("unable to parse feed dao id")
+	}
+
+	var readAt *common.Time
+	if item.ReadAt != nil {
+		readAt = common.NewTime(item.ReadAt.AsTime())
+	}
+
+	var archivedAt *common.Time
+	if item.ArchivedAt != nil {
+		archivedAt = common.NewTime(item.ArchivedAt.AsTime())
+	}
+
+	return feed.Item{
+		ID:           feedID,
+		CreatedAt:    *common.NewTime(item.CreatedAt.AsTime()),
+		UpdatedAt:    *common.NewTime(item.UpdatedAt.AsTime()),
+		ReadAt:       readAt,
+		ArchivedAt:   archivedAt,
+		DaoID:        daoID,
+		ProposalID:   item.GetProposalId(),
+		DiscussionID: item.GetDiscussionId(),
+		Type:         item.GetType(),
+		Action:       item.GetAction(),
+		DAO:          daoItem,
+		Proposal:     proposalItem,
+		Timeline:     convertFeedTimelineToInternal(item.Timeline),
+	}
 }